+++ pip_services_commons/config/JsonConfigReader.py
@@ -22,7 +22,7 @@ class JsonConfigReader(FileConfigReader):
     def __init__(self, path):
         super(JsonConfigReader, self).__init__(path)
 
-    def _read_object(self, correlation_id):
+    def _read_object(self, correlation_id, parameters):
         path = self.get_path()
         if path == None:
             raise ConfigException(correlation_id, "NO_PATH", "Missing config file path")
@@ -32,7 +32,9 @@ class JsonConfigReader(FileConfigReader):
         
         try:
             with open(path, 'r') as file:
-                return json.load(file)
+                config = file.read()
+                config = self._parameterize(config, parameters)
+                return json.loads(config)
         except Exception as ex:
             raise FileException(
                 correlation_id,
@@ -40,15 +42,15 @@ class JsonConfigReader(FileConfigReader):
                 "Failed reading configuration " + path + ": " + str(ex)
             ).with_details("path", path).with_cause(ex)
 
-    def _perform_read_config(correlation_id):
-        value = self._read_object(correlation_id)
+    def read_config(correlation_id, parameters):
+        value = self._read_object(correlation_id, parameters)
         return ConfigParams.from_value(value)
 
     @staticmethod
-    def read_object(correlation_id, path):
-        return JsonConfigReader(path)._read_object(correlation_id)
+    def read_object(correlation_id, path, parameters):
+        return JsonConfigReader(path)._read_object(correlation_id, parameters)
 
     @staticmethod
-    def read_config(correlation_id, path):
-        value = JsonConfigReader(path)._read_object(correlation_id)
+    def read_config(correlation_id, path, parameters):
+        value = JsonConfigReader(path)._read_object(correlation_id, parameters)
         return ConfigParams.from_value(value)