+++ test/config/test_JsonConfigReader.py
@@ -10,13 +10,18 @@
 import pytest
 
 from pip_services_commons.config import JsonConfigReader
+from pip_services_commons.config import ConfigParams
 
 class TestJsonConfigReader:
 
     def test_read_config(self):
-        config = JsonConfigReader.read_config(None, "data/config.json")
+        parameters = ConfigParams.from_tuples(
+            "param1", "Test Param 1",
+            "param2", "Test Param 2"
+        )
+        config = JsonConfigReader.read_config(None, "data/config.json", parameters)
         
-        assert 7 == len(config)
+        assert 9 == len(config)
         assert 123 == config.get_as_integer("Field1.Field11")
         assert "ABC" == config.get_as_string("Field1.Field12")
         assert 123 == config.get_as_integer("Field2.0")
@@ -24,3 +29,7 @@ class TestJsonConfigReader:
         assert 543 == config.get_as_integer("Field2.2.Field21")
         assert "XYZ" == config.get_as_string("Field2.2.Field22")
         assert True == config.get_as_boolean("Field3")
+        assert "XYZ" == config.get_as_string("Field2.2.Field22")
+        assert "Test Param 1" == config.get_as_string("Field4")
+        assert "Test Param 2" == config.get_as_string("Field5")
+